docker_compose/empty_project/app/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.13.0
django-split-settings==1.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
flake8==6.0.0
gunicorn==20.1.0
isort==5.12.0
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
uWSGI==2.0.21
